+++ src/main/java/org/accounting/system/endpoints/SystemAdminEndpoint.java
@@ -3,13 +3,18 @@ package org.accounting.system.endpoints;
 import io.quarkus.security.Authenticated;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DefaultValue;
+import jakarta.ws.rs.GET;
 import jakarta.ws.rs.PATCH;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
@@ -18,14 +23,19 @@ import org.accounting.system.beans.RequestUserContext;
 import org.accounting.system.constraints.NotFoundEntity;
 import org.accounting.system.dtos.InformativeResponse;
 import org.accounting.system.dtos.admin.ProjectRegistrationRequest;
+import org.accounting.system.dtos.authorization.request.AssignRoleRequestDto;
+import org.accounting.system.dtos.authorization.request.DetachRoleRequestDto;
+import org.accounting.system.dtos.client.ClientResponseDto;
 import org.accounting.system.dtos.project.ProjectRequest;
 import org.accounting.system.dtos.project.UpdateProjectRequest;
 import org.accounting.system.dtos.resource.ResourceRequest;
 import org.accounting.system.dtos.resource.ResourceResponse;
 import org.accounting.system.entities.projections.normal.ProjectProjection;
 import org.accounting.system.interceptors.annotations.SystemAdmin;
+import org.accounting.system.repositories.client.ClientRepository;
 import org.accounting.system.repositories.project.ProjectRepository;
 import org.accounting.system.services.SystemAdminService;
+import org.accounting.system.services.client.ClientService;
 import org.accounting.system.util.AccountingUriInfo;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.eclipse.microprofile.openapi.annotations.Operation;
@@ -40,6 +50,8 @@ import org.eclipse.microprofile.openapi.annotations.security.SecurityRequirement
 import org.eclipse.microprofile.openapi.annotations.security.SecurityScheme;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
+import static org.eclipse.microprofile.openapi.annotations.enums.ParameterIn.QUERY;
+
 @Path("/admin")
 @Authenticated
 @SecurityScheme(securitySchemeName = "Authentication",
@@ -50,6 +62,9 @@ import org.eclipse.microprofile.openapi.annotations.tags.Tag;
         in = SecuritySchemeIn.HEADER)
 
 public class SystemAdminEndpoint {
+
+    @Inject
+    ClientService clientService;
 
     @Inject
     SystemAdminService systemAdminService;
@@ -287,6 +302,179 @@ public class SystemAdminEndpoint {
             @PathParam("id") @Valid @NotFoundEntity(repository = ProjectRepository.class, id = String.class, message = "There is no Project with the following id:") String id, @Valid @NotNull(message = "The request body is empty.") UpdateProjectRequest updateProjectRequest) {
 
         var response = systemAdminService.updateProject(updateProjectRequest, id);
+
+        return Response.ok().entity(response).build();
+    }
+
+    @Tag(name = "System Administrator")
+    @Operation(
+            summary = "Returns the available clients.",
+            description = "This operation fetches the registered Accounting System clients. By default, the first page of 10 Clients will be returned. " +
+                    "You can tune the default values by using the query parameters page and size.")
+    @APIResponse(
+            responseCode = "200",
+            description = "Array of available clients.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = ClientEndpoint.PageableClientResponseDto.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Errors.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+
+    @GET
+    @Path("/clients")
+    @Produces(value = MediaType.APPLICATION_JSON)
+    @SystemAdmin
+    public Response getClients(@Parameter(name = "page", in = QUERY,
+            description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
+                               @Parameter(name = "size", in = QUERY,
+                                       description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                               @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                               @Context UriInfo uriInfo){
+
+        var serverInfo = new AccountingUriInfo(serverUrl.concat(basePath).concat(uriInfo.getPath()));
+
+        return Response.ok().entity(clientService.findAllClientsPageable(page-1, size, serverInfo)).build();
+    }
+
+    @Tag(name = "System Administrator")
+    @Operation(
+            hidden = true,
+            summary = "Assign one or more roles to a registered client.",
+            description = "Using the unique identifier of a registered client, you can assign roles to it.")
+    @APIResponse(
+            responseCode = "200",
+            description = "The Roles have been successfully assigned to a registered client.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = ClientResponseDto.class)))
+    @APIResponse(
+            responseCode = "400",
+            description = "Bad Request.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "Client has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "The authenticated client is not permitted to perform the requested operation.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Role not found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "415",
+            description = "Cannot consume content type.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Errors.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+
+    @POST
+    @Path("clients/{client_id}/assign-roles")
+    @Produces(value = MediaType.APPLICATION_JSON)
+    @Consumes(value = MediaType.APPLICATION_JSON)
+    @SystemAdmin
+    public Response assignRole(@Valid @NotNull(message = "The request body is empty.") AssignRoleRequestDto assignRoleRequestDto,
+                               @Parameter(
+                                       description = "client_id is the unique identifier of a client.",
+                                       required = true,
+                                       example = "xyz@example.org",
+                                       schema = @Schema(type = SchemaType.STRING))
+                               @PathParam("client_id")
+                               @Valid
+                               @NotFoundEntity(repository = ClientRepository.class, id = String.class, message = "There is no registered Client with the following id:") String clientId){
+
+        var response = clientService.assignRolesToRegisteredClient(clientId, assignRoleRequestDto.roles);
+
+        return Response.ok().entity(response).build();
+    }
+
+    @Tag(name = "System Administrator")
+    @Operation(
+            hidden = true,
+            summary = "Detach one or more roles from a registered client.",
+            description = "Using the unique identifier of a registered client, you can detach roles from it.")
+    @APIResponse(
+            responseCode = "200",
+            description = "The Roles have been successfully detached from a registered client.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = ClientResponseDto.class)))
+    @APIResponse(
+            responseCode = "400",
+            description = "Bad Request.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "Client has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "The authenticated client is not permitted to perform the requested operation.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Role not found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "415",
+            description = "Cannot consume content type.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Errors.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+
+    @POST
+    @Path("clients/{client_id}/detach-roles")
+    @Produces(value = MediaType.APPLICATION_JSON)
+    @Consumes(value = MediaType.APPLICATION_JSON)
+    @SystemAdmin
+    public Response detachRole(@Valid @NotNull(message = "The request body is empty.") DetachRoleRequestDto detachRoleRequestDto,
+                               @Parameter(
+                                       description = "client_id is the unique identifier of a client.",
+                                       required = true,
+                                       example = "xyz@example.org",
+                                       schema = @Schema(type = SchemaType.STRING))
+                               @PathParam("client_id")
+                               @Valid
+                               @NotFoundEntity(repository = ClientRepository.class, id = String.class, message = "There is no registered Client with the following id:") String clientId){
+
+        var response = clientService.detachRolesFromRegisteredClient(clientId, detachRoleRequestDto.roles);
 
         return Response.ok().entity(response).build();
     }